Fix for buffer overflow
This is a somewhat naive fix for a possible buffer overflow when handling
packets that span more than one packet.  The existing codebase doesn't
do any sanity checking that the packet is within the required bounds
of the buffer (see DEFBUF), so this really should be a part of a more
robust handling mechanism.  However, this is sufficient for now.