I have another Idea
If we just hook directily into the audioplayer (usbm or others I'm not thying to make this too heavy ) using the PlayResumeData() method to pause/play it.  I was trying to avoid getting into specific modules so if you can think of a more general one to hook into that would be greeaaat.